requirements.txt
fastapi
uvicorn
python-dotenv
openai
pinecone
sse-starlette
